+++ MLEM.Ui/Elements/Paragraph.cs
@@ -110,19 +110,15 @@ namespace MLEM.Ui.Elements {
         /// <param name="width">The paragraph's width. Note that its height is automatically calculated.</param>
         /// <param name="textCallback">The paragraph's text</param>
         /// <param name="autoAdjustWidth">Whether the paragraph's width should automatically be calculated based on the text within it.</param>
-        public Paragraph(Anchor anchor, float width, TextCallback textCallback, bool autoAdjustWidth = false)
-            : this(anchor, width, "", autoAdjustWidth) {
+        public Paragraph(Anchor anchor, float width, TextCallback textCallback, bool autoAdjustWidth = false) : this(anchor, width, "", autoAdjustWidth) {
+            this.IsHidden = true;
             this.GetTextCallback = textCallback;
-            this.Text = textCallback(this);
-            if (this.Text == null)
-                this.IsHidden = true;
         }
 
         /// <inheritdoc cref="Paragraph(Anchor,float,TextCallback,bool)"/>
         public Paragraph(Anchor anchor, float width, string text, bool autoAdjustWidth = false) : base(anchor, new Vector2(width, 0)) {
+            this.IsHidden = true;
             this.Text = text;
-            if (this.Text == null)
-                this.IsHidden = true;
             this.AutoAdjustWidth = autoAdjustWidth;
             this.CanBeSelected = false;
             this.CanBeMoused = false;